BitKeeper/etc/ignore:
  auto-union
BitKeeper/etc/logging_ok:
  auto-union
acconfig.h:
  Auto merged
acinclude.m4:
  Auto merged
include/myisam.h:
  Auto merged
mysql-test/mysql-test-run.sh:
  Auto merged
BitKeeper/deleted/.del-ib_config.h.in~9e57db8504e55b7:
  Auto merged
BitKeeper/deleted/.del-ib_config.h~7539e26ffc614439:
  Auto merged
client/mysqltest.c:
  Auto merged
sql/lex.h:
  Auto merged
sql/sql_base.cc:
  Auto merged
sql/sql_parse.cc:
  Auto merged
sql/sql_select.h:
  Auto merged
sql/sql_show.cc:
  Auto merged
sql/table.h:
  Auto merged
sql/mysql_priv.h:
  Auto merged
sql/sql_class.h:
  Auto merged
Docs/manual.texi:
  merged
client/errmsg.c:
  merged
configure.in:
  merged
sql/mysqld.cc:
  merged
sql/sql_select.cc:
  merged, needs manual fixing
sql/sql_yacc.yy:
  merged, needs manual fixing

+++ extra/resolve_stack_dump.c
@@ -303,7 +303,7 @@ static void do_resolve()
 	  uchar* addr = (uchar*)read_addr(&p);
 	  if(resolve_addr(addr, &se))
 	    fprintf(fp_out, "%p %s + %d\n", addr, se.symbol,
-		    addr - se.addr);
+		    (int) (addr - se.addr));
 	  else
 	    fprintf(fp_out, "%p (?)\n", addr);
